atts

+++ projeto/src/controler/cliente_bd/excluirCliente.php
@@ -1,14 +1,20 @@
 <?php
+require_once "../../protect.php";
 require_once "../../conexao.php";
 
-if(!isset($_SESSION)){
-    session_start();
-}
-$idCliente = isset($_POST['idcliente']) ? $_POST['idcliente'] : 0;
-
-$del = "DELETE FROM cliente WHERE idcliente = '$idCliente' "; 
-$sql_query = $conexao->query($sql_code);
-// $delgo = sql_query ($del) or die ('Erro ao deletar'); 
-// echo "deletado";
-
-header("../../index.php");
+    if(!isset($_SESSION['tipo']) && $_SESSION['tipo'] == "Administrador"){
+        header("Location: nao_permitido.php");
+        
+        $idCliente = isset($_GET['id']) ? $_GET['id'] : 0;
+        
+        if($idCliente > 0){
+            $del = "DELETE FROM cliente WHERE idcliente = '$idCliente' "; 
+            $sql_query = $conexao->query($del);
+            
+            
+        }
+        
+        header("../../../clientes.php");
+    } else {
+        header("Location: ../../../nao_permitido.php");
+    }
